+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix reimport via query (IndexError bugfix) <https://github.com/gtronset/beets-filetote/pull/126>
+
 ## [0.4.5] - 2023-12-01
 
 ### Added